# openapi.yaml
openapi: 3.0.3
info:
  title: Gateway Zero API
  description: |
    Gateway Zero 是一个基于 Rust 和 Axum 框架开发的 AI API 网关系统。
    它提供统一的 OpenAI 兼容接口，支持多个 AI 提供商（OpenAI、Anthropic、智谱 AI 等），
    并具备负载均衡、令牌管理、用量统计等企业级功能。
  version: 0.1.0
  contact:
    name: Gateway Zero Team
servers:
  - url: http://localhost:8080
    description: 本地开发服务器
  - url: https://api.example.com
    description: 生产环境服务器
components:
  securitySchemes:
    # Client Token：外部调用 /v1/* 的 API Token（与 Admin Identity 严格区分）
    ClientToken:
      type: apiKey
      in: header
      name: Authorization
      description: "Client Token 认证（格式: Bearer <token>），用于访问 `/v1/*`"

    # Admin Identity：管理员身份（可来自 JWT AccessToken / TUI Session bearer / Web Session Cookie）
    AdminTuiSessionToken:
      type: http
      scheme: bearer
      description: 管理员身份（Admin Identity）的 TUI Session Token（bearer，但非 JWT）

    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: gw_session
      description: 管理员身份（Admin Identity）的 Web Session Cookie（HttpOnly）

    AccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "AccessToken（JWT），格式: Bearer <jwt>；用于访问 `/auth/me`、`/auth/change-password`、`/me/*` 以及 superadmin-only 的 `/admin/*`、`/providers/*`"
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: 错误代码 (如 http_error, not_found, rate_limited 等)
        message:
          type: string
          description: 错误详细信息

    # 请求日志
...
    Provider:
      type: object
      properties:
        name:
          type: string
          description: 提供商名称
        api_type:
          type: string
          enum: [openai, anthropic, zhipu]
          description: 提供商类型
        base_url:
          type: string
          description: API 基础 URL
        api_keys:
          type: array
          items:
            type: string
          description: API Keys 列表
        models_endpoint:
          type: string
          nullable: true
          description: 模型列表端点
        enabled:
          type: boolean
          description: 是否启用（禁用后调度不会选择该提供商）
      required:
        - name
        - api_type
        - base_url
        - api_keys
        - enabled
...
    ChatCompletionRequest:
      type: object
      properties:
        model:
          type: string
          description: "模型名称 (格式: provider/model 或 model)"
        messages:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [system, user, assistant]
              content:
                type: string
          description: 消息列表
        stream:
          type: boolean
          default: false
          description: 是否使用流式响应
        temperature:
          type: number
          format: float
          nullable: true
          description: 温度参数
        max_tokens:
          type: integer
          nullable: true
          description: 最大生成 tokens
        top_p:
          type: number
          format: float
          nullable: true
          description: Top-p 采样参数
      required:
        - model
        - messages

    # 聊天补全响应
    ChatCompletionResponse:
      type: object
      properties:
        id:
          type: string
          description: 响应ID
        object:
          type: string
          default: chat.completion
        created:
          type: integer
          format: int64
          description: 创建时间戳
        model:
          type: string
          description: 使用的模型
        choices:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              message:
                type: object
                properties:
                  role:
                    type: string
                  content:
                    type: string
              finish_reason:
                type: string
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer

    # 令牌余额响应（/v1/token/balance）
...
paths:
...
  /auth/keys/{fingerprint}:
    delete:
      summary: 删除管理员公钥
      description: 管理员身份接口（Admin Identity）
      operationId: deleteAdminKey
      tags:
        - Auth
      security:
        - AccessToken: []
        - AdminTuiSessionToken: []
        - AdminSessionCookie: []
      parameters:
        - name: fingerprint
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/chat/completions:
    post:
      summary: 聊天补全
      description: 创建聊天补全请求，支持流式和非流式响应
      operationId: createChatCompletion
      tags:
        - Chat
      security:
        - ClientToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
            text/event-stream:
              schema:
                type: string
                description: SSE 流式响应
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /providers/{provider}:
    get:
      summary: 获取提供商详情
      description: 获取指定提供商的详细信息
      operationId: getProvider
      tags:
        - Providers
      security:
        - AccessToken: []
        - AdminTuiSessionToken: []
        - AdminSessionCookie: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
          description: 提供商名称
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 提供商不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: 更新提供商
      description: 更新指定提供商的配置
      operationId: updateProvider
      tags:
        - Providers
      security:
        - AccessToken: []
        - AdminTuiSessionToken: []
        - AdminSessionCookie: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
          description: 提供商名称
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                api_type:
                  type: string
                  enum: [openai, anthropic, zhipu]
                base_url:
                  type: string
                models_endpoint:
                  type: string
                  nullable: true
              required:
                - api_type
                - base_url
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 提供商不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: 删除提供商
      description: 删除指定的提供商配置
      operationId: deleteProvider
      tags:
        - Providers
      security:
        - AccessToken: []
        - AdminTuiSessionToken: []
        - AdminSessionCookie: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
          description: 提供商名称
      responses:
        '200':
          description: 成功删除
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 提供商不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/providers/{provider}/keys/stats:
    get:
      summary: 获取提供商密钥统计（按统计口径）
      description: |
        返回该提供商下各密钥在指定统计口径内的请求统计（总请求/成功/失败/可用率）。

        优先级（从高到低）：
        1) `range`
        2) `start_date`/`end_date`
        3) `window_minutes`
      operationId: getAdminProviderKeysStats
      tags:
        - Providers
      security:
        - AccessToken: []
        - AdminTuiSessionToken: []
        - AdminSessionCookie: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
          description: 提供商名称
        - name: range
          in: query
          schema:
            type: string
            enum: [hour, day, week, since_created]
          description: 统计口径（可选；优先级最高）
        - name: window_minutes
          in: query
          schema:
            type: integer
            default: 60
            minimum: 1
            maximum: 10080
          description: 统计窗口（分钟；未指定 range/start_date/end_date 时生效）
        - name: start_date
          in: query
          schema:
            type: string
          description: 开始日期（YYYY-MM-DD，可选）
        - name: end_date
          in: query
          schema:
            type: string
          description: 结束日期（YYYY-MM-DD，可选）
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                type: object
                properties:
                  window_minutes:
                    type: integer
                    format: int64
                    description: 统计窗口（分钟；range=since_created 时为 0）
                  start_date:
                    type: string
                    nullable: true
                  end_date:
                    type: string
                    nullable: true
                  generated_at:
                    type: string
                    format: date-time
                  keys:
                    type: array
                    items:
                      type: object
                      properties:
                        masked_key:
                          type: string
                          description: 脱敏密钥（4+****+4）
                        total_requests:
                          type: integer
                          format: int64
                        success_count:
                          type: integer
                          format: int64
                        failure_count:
                          type: integer
                          format: int64
                        availability_rate:
                          type: integer
                          format: int32
                          minimum: 0
                          maximum: 100
                      required: [masked_key, total_requests, success_count, failure_count, availability_rate]
                required: [window_minutes, keys, generated_at]
              examples:
                hour:
                  summary: 近 1 小时
                  value:
                    window_minutes: 60
                    keys:
                      - masked_key: "sk-a***1234"
                        total_requests: 12
                        success_count: 10
                        failure_count: 2
                        availability_rate: 83
                    generated_at: "2025-01-01T00:00:00Z"
                since_created:
                  summary: 自创建以来
                  value:
                    window_minutes: 0
                    keys:
                      - masked_key: "sk-a***1234"
                        total_requests: 120
                        success_count: 110
                        failure_count: 10
                        availability_rate: 92
                    generated_at: "2025-01-01T00:00:00Z"
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 提供商不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...